requirements.txt
mxnet-cu80==1.3.1
numpy==1.14.6
opencv-python==4.0.0
